Support raw, headerless image output
This set of changes removes the `packed` flag, which should be the default behaviour, and replaced it with a `raw` flag which, when supplied, causes the image packing tool to output raw RGB or RGBA values with no header. Raw image data can be used to initalise a surface and should not be loaded with `SpriteSheet::load`.