Make all formatters have components
Formatters which only generate a single file and use the supplied
filename directly should have components=(None, ). The OutputFormat
decorator sets this value by default if no components are supplied.

This also makes join() functions handle all components at the same
and return a dict, the same way fragments() functions do.

This will allow the removal of a lot of special-case code that tries
to handle the different return types of formatters.