Implement AssetWriter
This class represents the process of turning a collection of assets
into a set of output files. To use it, you create an instance, add
(symbol, data) pairs, and then when you are finished, call write().

This replaces the Tool.output() method, and the entire AssetTarget
class. It also handles guessing the output format from the path.

It has some extra features the others don't. It can sort the assets
before writing, and it can generate a report detailing the assets
and files it generated.