Split out RLE code into functions and re-implement with a generator
This splits the RLE code into two functions. One to count repetitions
and the other to build the RLE'd bytes. Note that the repeat count
includes the initial instance, ie 1 means the value appeared once.
This is different to the binary representation, where repeat count
of 1 means the value appears twice in the output. This is necessary
in order to make the for loop work.