Auto-register all Tools and load them dynamically
This makes all Tools get loaded dynamically, without having to load
individually import each one and add them to the argparse machinery.

The typemap for builders is extended with a boolean value for each
filename extension. If true, the builder is registered as the default
handler for this extension. This is to prevent clashes where multiple
builders can handle the same filetype, eg tiles and raw.