Fix the raw_binary joiner
This function slipped through the refactoring because it isn't used
in any of the tests or examples. Rename the argument to "fragments"
for consistency, and fetch the data to be joined from the None key
because this joiner only produces one output file and doesn't add a
suffix to it.