Refactor image packing code into a Construct Adapter
This allows using parse() and build() directly on 8 bit indexed
pixels, without having to worry about the packing format. Just set
type to None or leave it out entirely and ImageCompressor will
select the best mode. Or specify a mode, and ImageCompressor will
use it.

In order to make this work the image struct had to be cleaned up a
little, and the parsed object representation is a bit different.
Most of the information is now nested under 'data' which calculates
the size field internally (and it is no longer exposed) and the
thing that used to be called 'data' is now 'data.pixels'.

This also brings together code that was spread around into a single
module, core.compression.

Also adds tests specifically targeting the parsing code.