Add install command and help for flash commands
This is a smarter version of `flash flash` and `flash sd`. If no
destination is specified, .blit files are written to flash, and
all others are written to the root of the SD card. If a destination
is specified, then all files are written to that directory on the
SD card.