Support for 16bit tmx tile indexes, replaces #74
Adds support for TMX tile indexes that are greater than 255 tiles, this means 8x8 tile size sprite sheets greater than 128x128 in size can be used and the right tile IDs will be output, rather than throwing an error.

The data output is still uint8_t so a `TMX` struct that treats the layer data as `uint16_t` is required.

16bit tile data is currently for **advanced** users only, and will not work with `Map` or `TileMap` which both assume 8bit indexes.