Bail if no valid input files found, for #66
This change will cause the asset builder to bail early if no valid input files are found.

Files are filtered based on whether they actually exist and the target will bail early if it has missing sources.

Note: since glob is open-ended it will fail if it finds nothing, but can't offer any guarantee it'll find what a user expects.